rename dir contracts to projects
