+++ apps/docs-snippets/projects/counter/src/main.sw
