+++ apps/docs-snippets/projects/echo-configurables/src/main.sw
