+++ apps/docs-snippets/projects/echo-values/src/main.sw
