+++ apps/docs-snippets/projects/index.ts
