+++ apps/docs-snippets/projects/log-values/src/main.sw
