+++ apps/docs-snippets/projects/return-context/src/main.sw
