+++ apps/docs-snippets/projects/simple-token-abi/Forc.toml
