+++ apps/docs-snippets/projects/simple-token-abi/src/main.sw
