+++ apps/docs-snippets/projects/simple-token/Forc.toml
