+++ apps/docs-snippets/projects/simple-token/src/main.sw
