+++ apps/docs-snippets/projects/token-depositor/Forc.toml
