+++ apps/docs-snippets/projects/token-depositor/src/main.sw
