+++ apps/docs-snippets/projects/whitelisted-address-predicate/Forc.toml
