+++ apps/docs-snippets/projects/whitelisted-address-predicate/src/main.sw
