+++ apps/docs-snippets/src/guide/contracts/deploying-contracts.test.ts
@@ -2,7 +2,7 @@ import { readFileSync } from 'fs';
 import { Provider, FUEL_NETWORK_URL, Wallet, ContractFactory } from 'fuels';
 import { join } from 'path';
 
-import { SnippetContractEnum } from '../../../contracts';
+import { SnippetContractEnum } from '../../../projects';
 import { getTestWallet } from '../../utils';
 
 describe(__filename, () => {
