+++ apps/docs-snippets/src/guide/contracts/transaction-parameters.test.ts
@@ -1,7 +1,7 @@
 import type { Contract } from 'fuels';
 import { BN, ContractFactory } from 'fuels';
 
-import { SnippetContractEnum, getSnippetContractArtifacts } from '../../../contracts';
+import { SnippetContractEnum, getSnippetContractArtifacts } from '../../../projects';
 import { getTestWallet } from '../../utils';
 
 describe(__filename, () => {
