+++ apps/docs/src/guide/contracts/logs.md
@@ -4,7 +4,7 @@ When you log a value within a contract method, it generates a log entry that is 
 
 Consider the following example contract:
 
-<<< @/../../docs-snippets/contracts/log-values/src/main.sw#log-1{rust:line-numbers}
+<<< @/../../docs-snippets/projects/log-values/src/main.sw#log-1{rust:line-numbers}
 
 To access the logged values in TypeScript, use the `logs` property in the `FunctionInvocationResult` of a contract call result. The logs data will be stored in an `Array<any>`:
